Distribute real es module entries
