+++ .babelrc
@@ -1,0 +1,9 @@
+{
+  "presets": [
+    ["@babel/env", {
+      loose: true,
+      modules: false,
+      forceAllTransforms: true
+    }]
+  ]
+}
