+++ .gitignore
@@ -1,3 +1,6 @@
+# Built bundles
+dist
+
 # Logs
 logs
 *.log
