--- .npmignore
@@ -1,7 +1,0 @@
-/.*/
-/.*
-/benchmarks/
-/tests/
-/bower.json
-/CONTRIBUTING.md
-/HISTORY.md
