+++ package.json
@@ -2,18 +2,28 @@
   "name": "classnames",
   "version": "2.2.5",
   "description": "A simple utility for conditionally joining classNames together",
-  "main": "index.js",
+  "main": "dist/classnames.cjs.js",
+  "module": "dist/classnames.es.js",
   "author": "Jed Watson",
   "license": "MIT",
   "repository": {
     "type": "git",
     "url": "https://github.com/JedWatson/classnames.git"
   },
+  "files": [
+    "dist",
+    "bind",
+    "dedupe"
+  ],
   "scripts": {
+    "prebuild": "rimraf dist",
+    "build": "rollup -c",
     "benchmarks": "node ./benchmarks/run",
     "benchmarks-browserify": "./node_modules/.bin/browserify ./benchmarks/runInBrowser.js >./benchmarks/runInBrowser.bundle.js",
     "benchmarks-in-browser": "./node_modules/.bin/opn ./benchmarks/benchmarks.html",
-    "test": "mocha tests/*.js"
+    "pretest": "npm run build",
+    "test": "mocha tests/*.js",
+    "prepublish": "npm test && npm run build"
   },
   "keywords": [
     "react",
@@ -25,9 +35,14 @@
     "utility"
   ],
   "devDependencies": {
+    "@babel/core": "7.0.0-beta.35",
+    "@babel/preset-env": "7.0.0-beta.35",
     "benchmark": "^1.0.0",
     "browserify": "^14.1.0",
     "mocha": "^2.1.0",
-    "opn-cli": "^3.1.0"
+    "opn-cli": "^3.1.0",
+    "rimraf": "^2.6.2",
+    "rollup": "^0.52.3",
+    "rollup-plugin-babel": "4.0.0-beta.0"
   }
 }
